+++ src/mpid/ch4/netmod/ofi/netmod_inline.h
@@ -9,6 +9,7 @@
 #include "ofi_am.h"
 #include "ofi_events.h"
 #include "ofi_proc.h"
+#include "ofi_progress.h"
 #include "ofi_unimpl.h"
 
 #ifdef MPIDI_ENABLE_AM_ONLY
