+++ src/mpid/ch4/netmod/ofi/ofi_progress.c
@@ -7,9 +7,7 @@
 #include "ofi_impl.h"
 #include "ofi_events.h"
 
-static int handle_deferred_ops(void);
-
-static int handle_deferred_ops(void)
+int handle_deferred_ops(void)
 {
 
     int mpi_errno = MPI_SUCCESS;
@@ -48,42 +46,4 @@ static int handle_deferred_ops(void)
     return mpi_errno;
   fn_fail:
     goto fn_exit;
-}
-
-int MPIDI_OFI_progress(int vci, int blocking)
-{
-    int mpi_errno;
-    struct fi_cq_tagged_entry wc[MPIDI_OFI_NUM_CQ_ENTRIES];
-    ssize_t ret;
-    MPIR_FUNC_VERBOSE_STATE_DECL(MPID_STATE_MPIDI_OFI_PROGRESS);
-    MPIR_FUNC_VERBOSE_ENTER(MPID_STATE_MPIDI_OFI_PROGRESS);
-
-    int vni = MPIDI_OFI_vci_to_vni(vci);
-    if (vni >= MPIDI_OFI_global.num_vnis) {
-        /* The requests generated from ofi will have vni/vci within our range.
-         * All requests that have vci beyond num_vnis are not from us, nothing
-         * to do, so simply return.
-         * NOTE: it is not an error since global progress will poll every vci.
-         */
-        return MPI_SUCCESS;
-    }
-
-    if (unlikely(MPIDI_OFI_get_buffered(wc, 1)))
-        mpi_errno = MPIDI_OFI_handle_cq_entries(wc, 1);
-    else if (likely(1)) {
-        ret = fi_cq_read(MPIDI_OFI_global.ctx[vni].cq, (void *) wc, MPIDI_OFI_NUM_CQ_ENTRIES);
-
-        if (likely(ret > 0))
-            mpi_errno = MPIDI_OFI_handle_cq_entries(wc, ret);
-        else if (ret == -FI_EAGAIN)
-            mpi_errno = MPI_SUCCESS;
-        else
-            mpi_errno = MPIDI_OFI_handle_cq_error(vni, ret);
-    }
-
-    handle_deferred_ops();
-
-    MPIR_FUNC_VERBOSE_EXIT(MPID_STATE_MPIDI_OFI_PROGRESS);
-
-    return mpi_errno;
 }
