+++ src/mpid/ch4/netmod/ucx/netmod_inline.h
@@ -18,6 +18,7 @@
 #include "ucx_part.h"
 #endif
 
+#include "ucx_progress.h"
 #include "ucx_request.h"
 #include "ucx_am.h"
 #include "ucx_proc.h"
